Made the animatronics smaller and more quirky
Before this change they were like double the size of the hallways
between rooms and it was just hard to get them to fit in places that
they were too big to fit in. So I had to make them smaller so they could
fit in small places. For example the office.