Made fredrick move
Right now he will always make it in your office and turn off your doors
but at some point I will make it so you can flash your light at him
while he is in the hallway to make him go back to the dining room. Also
can we talk about how adorable he is when he is sitting in the office?
idk why but I find that so cute